test: unit test ante handler
